# src/main/java/wit/edu/inz/ticket/entity/TicketType.java
package wit.edu.inz.ticket.entity;

public enum TicketType {
    WATER_LEAK,
    CLOGGED_DRAIN,
    HEATING_FAILURE,
    POWER_OUTAGE,
    LIGHT_FIXTURE,
    BROKEN_WINDOW,
    BROKEN_DOOR,
    APPLIANCE_REPAIR,
    CLEANING_REQUEST,
    OTHER
}