# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Maintenance Request Manager
  version: 1.0.0
tags:
  - name: Authentication
    description: User registration and authentication

  - name: Profile
    description: Operations related to the authenticated user's profile

  - name: Ticket - User Operations
    description: Ticket operations available to residents

  - name: Ticket - Worker Operations
    description: Ticket operations available to maintenance workers

  - name: Comments
    description: Operations related to ticket comments

  - name: Attachments
    description: Operations related to ticket attachments

  - name: Administration
    description: Administrative operations
paths:
  /register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
  /login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '401':
          description: Invalid username or password
  /tickets:
    get:
      tags:
        - Ticket - User Operations
      summary: Retrieve all tickets created by the authenticated user
      operationId: getTickets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
    post:
      tags:
        - Ticket - User Operations
      summary: Create a new ticket
      operationId: createTicket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
  /tickets/{id}:
    get:
      tags:
        - Ticket - User Operations
        - Ticket - Worker Operations
      summary: Retrieve a single ticket
      operationId: getTicket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          description: Ticket not found
    patch:
      tags:
        - Ticket - User Operations
      summary: Update ticket details
      operationId: updateTicket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdateRequest'
      responses:
        '200':
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
  /tickets/{id}/attachments:
    post:
      tags:
        - Attachments
      summary: Upload attachment to ticket
      operationId: createAttachment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Attachment uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentResponse'
    get:
      tags:
        - Attachments
      summary: Get attachments related to the ticket
      operationId: getAttachments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Fetched attachments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttachmentResponse'
  /tickets/attachments/{attachmentId}:
    get:
      tags:
        - Attachments
      summary: Download attachment
      operationId: downloadAttachment
      security:
        - bearerAuth: []
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Attachment file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /tickets/{id}/comments:
    post:
      tags:
        - Comments
      summary: Add a comment to a ticket
      operationId: createComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          description: Invalid comment data
        '401':
          description: Authentication required
        '404':
          description: Ticket not found
    get:
      tags:
        - Comments
      summary: Retrieve comments for a ticket
      operationId: getTicketComments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
        '401':
          description: Authentication required
        '404':
          description: Ticket not found
  /tickets/{id}/status:
    patch:
      tags:
        - Ticket - Worker Operations
      summary: Update the status of a ticket
      operationId: updateTicketStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketStatusUpdateRequest'
      responses:
        '200':
          description: Ticket status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          description: Invalid status provided
        '404':
          description: Ticket not found
  /tickets/{id}/priority:
    patch:
      tags:
        - Ticket - Worker Operations
      summary: Update the priority of a ticket
      operationId: updateTicketPriority
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdatePriorityRequest'
      responses:
        '200':
          description: Ticket priority updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          description: Invalid priority provided
        '404':
          description: Ticket not found
  /tickets/{id}/assign:
    patch:
      tags:
        - Ticket - Worker Operations
      summary: Assign the ticket to a maintenance worker
      operationId: assignTicket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Ticket response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          description: Invalid assignment
        '404':
          description: Ticket not found
  /admin/users:
    get:
      tags:
        - Administration
      summary: Retrieve all users
      operationId: getUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          description: Authentication required
        '403':
          description: Administrator role required
  /admin/tickets:
    get:
      tags:
        - Administration
      summary: Retrieve all tickets
      operationId: getAdminTickets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all tickets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
        '401':
          description: Authentication required
        '403':
          description: Administrator role required
  /admin/workers:
    post:
      tags:
        - Administration
      summary: Create a maintenance worker account
      operationId: createWorker
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerCreateRequest'
      responses:
        '201':
          description: Worker account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid worker data
        '401':
          description: Authentication required
        '403':
          description: Administrator role required
        '409':
          description: Username or email already exists
  /profile:
    get:
      tags:
        - Profile
      summary: Get authenticated user profile
      description: >
        Returns details of the currently authenticated user together with
        their currently open maintenance tickets.
      operationId: getProfile
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Unauthorized
        '404':
          description: User not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProfileResponse:
      type: object
      description: Details of the authenticated user and their open tickets
      required:
        - id
        - username
        - email
        - role
        - tickets
      properties:
        id:
          type: integer
          format: int64
          example: 3
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        username:
          type: string
          example: johnsmith
        email:
          type: string
          format: email
          example: john.smith@example.com
        role:
          $ref: '#/components/schemas/Role'
        tickets:
          type: array
          description: Tickets created by the user that are not completed or cancelled
          items:
            $ref: '#/components/schemas/ProfileTicketResponse'
    ProfileTicketResponse:
      type: object
      description: Lightweight ticket representation used on the profile page
      required:
        - id
        - title
        - status
      properties:
        id:
          type: integer
          format: int64
          example: 12

        title:
          type: string
          example: Leaking pipe in bathroom

        status:
          $ref: '#/components/schemas/TicketStatus'
    AuthenticationRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: john.doe
        password:
          type: string
          format: password
          example: myPassword123
      description: Authentication request containing user credentials.
    RegisterRequest:
      type: object
      required:
        - firstName
        - lastName
        - username
        - email
        - password
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: MySecurePassword123
        addressLine:
          type: string
          example: Main Street
        addressLine2:
          type: string
          example: Building A
        buildingNr:
          type: integer
          example: 15
        apartmentNr:
          type: integer
          example: 7
      description: Resident registration request.
    Role:
      type: string
      enum:
        - USER
        - WORKER
        - ADMIN
      description: Application role assigned to a user.
    AuthenticationResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiJ9...
      description: Authentication response containing the generated JWT access token.
    TicketCreateRequest:
      type: object
      required:
        - title
        - description
        - category
        - type
      properties:
        title:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/TicketCategory'
        type:
          $ref: '#/components/schemas/TicketType'
      description: Resident request for creating a maintenance ticket.
    TicketStatusUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/TicketStatus'
      description: Worker request for changing the status of a ticket.
    TicketUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/TicketCategory'
        type:
          $ref: '#/components/schemas/TicketType'
      description: Resident request for updating editable ticket details.
    TicketUpdatePriorityRequest:
      type: object
      required:
        - priority
      properties:
        priority:
          $ref: '#/components/schemas/TicketPriority'
      description: Worker request for changing the priority of a ticket.
    TicketResponse:
      type: object
      required:
        - id
        - title
        - description
        - category
        - type
        - status
        - priority
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Leaking pipe in bathroom
        description:
          type: string
          example: There is a constant leak underneath the bathroom sink.
        category:
          $ref: '#/components/schemas/TicketCategory'
        type:
          $ref: '#/components/schemas/TicketType'
        status:
          $ref: '#/components/schemas/TicketStatus'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        createdAt:
          type: string
          format: date-time
          example: '2026-08-19T14:30:00+02:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-08-19T15:45:00+02:00'
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: null
        createdById:
          type: integer
          format: int64
          example: 5
        assignedWorkerId:
          type: integer
          format: int64
          nullable: true
          example: 12
      description: Ticket information returned by ticket operations.
    TicketCategory:
      type: string
      enum:
        - PLUMBING
        - ELECTRICAL
        - HVAC
        - CARPENTRY
        - APPLIANCE
        - CLEANING
        - OTHER
      description: Category assigned to a maintenance ticket.
    TicketType:
      type: string
      enum:
        - WATER_LEAK
        - CLOGGED_DRAIN
        - HEATING_FAILURE
        - POWER_OUTAGE
        - LIGHT_FIXTURE
        - BROKEN_WINDOW
        - BROKEN_DOOR
        - APPLIANCE_REPAIR
        - CLEANING_REQUEST
        - OTHER
      description: Type of maintenance issue reported by the resident.
    TicketStatus:
      type: string
      enum:
        - OPEN
        - ASSIGNED
        - IN_PROGRESS
        - COMPLETED
        - CANCELLED
      description: Current lifecycle status of a maintenance ticket.
    TicketPriority:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH
        - URGENT
      description: Priority assigned to a maintenance ticket.
    TicketListResponse:
      type: array
      items:
        $ref: '#/components/schemas/TicketResponse'
      description: List of maintenance tickets.
    AttachmentResponse:
      type: object
      required:
        - id
        - filename
        - contentType
        - size
        - uploadedAt
      properties:
        id:
          type: integer
          format: int64
        filename:
          type: string
        contentType:
          type: string
        size:
          type: integer
          format: int64
        uploadedAt:
          type: string
          format: date-time
      description: Metadata describing a file attached to a ticket.
    CommentCreateRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Content of the comment
          example: The worker has been notified about the issue.
      description: Request for adding a comment to a ticket.
    CommentResponse:
      type: object
      required:
        - id
        - content
        - createdAt
        - ticketId
        - authorId
        - authorUsername
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        ticketId:
          type: integer
          format: int64
        authorId:
          type: integer
          format: int64
        authorUsername:
          type: string
      description: Comment information returned by comment operations.
    WorkerCreateRequest:
      type: object
      required:
        - firstName
        - lastName
        - username
        - email
        - password
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      description: Administrator request for creating a maintenance worker account.
    UserResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
        - username
        - email
        - role
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
      description: User information returned by administrative operations.
security:
  - bearerAuth: []